Cleanup leftovers from mcpClient
Move things around so we are using the now more modular approach.
